+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added event details in tooltip by hover (#217, #221)
 - Added option to Disable Event Time (#223)
+- Updated daily calendar view if timeRange changed (#229)
 
 ## 3.4.0 (2024-07-16)